pyproject.toml
[project]
name = "german-learning-cards"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
]
